backend/requirements.txt
requests
fastapi==0.103.2
uvicorn
pydantic==2.5.2
python-multipart
Pillow==9.5.0
pandas==2.0.2
opencv-python==4.7.0.72
pytesseract==0.3.10
matplotlib==3.8.3
imutils==0.5.4